Refactor code and ass testing file
